requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
google-generativeai==0.7.2
langgraph==0.0.66
langchain==0.2.16
langchain-google-genai==1.0.10
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
typing-extensions==4.8.0 